project/requirements.txt
Django==4.2.3
django-rest-framework==0.1.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.1.0